requirements.txt
azure-core==1.30.1
azure-identity==1.16.0
beautifulsoup4==4.12.3
boto3==1.34.131
chainlit==1.1.306
chromadb==0.4.24
html2text==2020.1.16
llama_index==0.10.58
llama-index-core==0.10.582.post1
llama-index-embeddings-bedrock==0.2.0
llama-index-llms-bedrock-converse==0.1.6
llama-index-postprocessor-cohere-rerank==0.1.7
llama-index-postprocessor-longllmlingua==0.1.2
llama-index-readers-web==0.1.18
llama-index-tools-google==0.1.5
llama-index-tools-tavily-research==0.1.3
llama-index-tools-wolfram-alpha==0.1.3
llama-index-vector-stores-qdrant==0.2.13
llmlingua==0.2.2
numpy==1.26.4
pandas==2.2.2
pyautogen==0.2.32
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qdrant-client==1.10.1
Requests==2.32.3
sec_api==1.0.18
ta==0.11.0
tavily-python==0.3.3
vanna==0.6.4
wolframalpha==5.1.3
yfinance==0.2.40
